back/requirements.txt
bleach==6.0.0
Django>=5.0,<5.1
django-anymail==12.0
django-cors-headers==3.5.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-hijack==3.4.1
django-model-utils==4.2.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.3.0
factory_boy==3.2.1
getconf==1.11.1
gunicorn==23.0.0
humanize==4.8.0
ipython==8.14.0
numpy==1.22.3
pandas==1.4.1
psycopg2==2.9.5
rollbar>=1.1,<1.2
telescoop_backup==0.1.4
wagtail>=6.1,<6.2
wagtail-generic-chooser==0.5.1
wagtail-localize==1.9
wagtail-modeladmin==2.0.0
wagtailsvg @ git+https://github.com/TelesCoop/wagtailsvg.git@wagtail6-support
